requirements.txt
# PDF处理和解析依赖
PyMuPDF==1.23.14          # PDF处理，转换为图片
pdf2image==1.17.0         # PDF转图片的替代方案
Pillow==10.2.0            # 图像处理



# OCR和文本识别
paddlepaddle-gpu==3.1.0         # 百度PaddleOCR，支持中英文OCR
easyocr==1.7.0            # EasyOCR，多语言OCR
pytesseract==0.3.10       # Tesseract OCR

# 版式分析
doclayout_yolo            # DocLayout YOLO模型，用于文档版式分析

# 表格解析
pdfplumber==0.10.0        # PDF表格提取
tabula-py==2.8.2          # 表格解析
pandas==2.2.0             # 数据处理

# 公式解析
sympy==1.12              # 符号数学库
matplotlib==3.8.2        # 绘图库
numpy==1.26.3            # 数值计算

# 机器学习和深度学习（用于版式分析）
torch==2.1.2             # PyTorch
torchvision==0.16.2      # 计算机视觉
transformers==4.37.0     # 预训练模型
ultralytics==8.0.196     # YOLO系列模型，用于版式分析
timm==0.9.12             # 深度学习模型库，LayoutReader依赖

# LayoutLMv3相关依赖（用于阅读顺序分析）
datasets==2.16.1         # 数据集处理
tokenizers==0.15.0       # 分词器
sentencepiece==0.1.99    # 文本处理

# 图像处理和计算机视觉
opencv-python<=4.6.0.66  # OpenCV
scikit-image==0.22.0     # 图像处理

# 文档处理和输出
markdown==3.5.2          # Markdown生成
jinja2==3.1.3           # 模板引擎
pyyaml==6.0.1            # YAML配置文件

# 工具库
loguru==0.7.2            # 日志记录
tqdm==4.66.1             # 进度条
click==8.1.7             # 命令行工具
pathlib2==2.3.7         # 路径处理

# Web界面
gradio==4.44.0           # Web界面构建
